if pt autorizare. doar la last resort facem asa.

+++ spring-app/src/main/java/victor/training/spring/web/controller/TrainingController.java
@@ -7,6 +7,7 @@ import org.owasp.html.Sanitizers;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.Resource;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import victor.training.spring.web.controller.dto.TrainingDto;
 import victor.training.spring.web.controller.dto.TrainingSearchCriteria;
@@ -61,6 +62,11 @@ public class TrainingController {
 	//  -> hasPermission + PermissionEvaluator [GEEK]
 	@DeleteMapping("{trainingId}")
 	public void delete(@PathVariable Long trainingId) {
+		var currentUser = SecurityContextHolder.getContext().getAuthentication().getName();
+		var roles = SecurityContextHolder.getContext().getAuthentication().getAuthorities(); // rolurile
+		if (roles.stream().noneMatch(r -> r.getAuthority().equals("ROLE_ADMIN"))) {
+			throw new RuntimeException("Only admins can delete");
+		}
 		trainingService.deleteById(trainingId);
 	}
